NIST.IR.8060 GEN-15 requires @version for File
The requirement is "include value if on exists for the file"
but swidval reports
ERROR GEN-15-2: The <File> element with the @name 'alias' has a blank @version attribute value.
if the attribute is not used or is set to empty string. To allow the
user to specify some value for the attribute when no one exists for
the file, adding command line parameter --default-file-version.