travis: Fix tests with Python 3.3
While the latest version of virtualenv used on Travis by default (16.0.0)
does support Python 3.3 the version of wheel it relies on (0.31.1) does
not anymore.

So use the previous version of virtualenv (15.2.0), which relies on wheel
0.30.0, with Python 3.3 to avoid this problem for now.

Interestingly, before 0.31.0 wheel didn't have any restrictions, even though
the changelog for 0.30.0 actually mentions the removal of support for Python
2.6, 3.2 and 3.3 (i.e. this might not actually work with Python 3.3,
however, we don't use wheel on Travis).