Remove root Directory tag without name attribute in hierarchic output
The name attribute is mandatory.  Instead of adding one (e.g. empty or
set to /) we remove the tag and set root="/" in the promoted Directory
tags if the path is absolute (that's also what's used for top-level
dirs in the flat output). Relative paths (like in the unit tests) don't
get a root attribute anymore (which seems to be different from the
behavior of the non-hierarchic file generator).  For files in the root
directory a root="/" attribute is added also if the path is absolute.

Fixes #50.