Remove support for Python 3.3, add support for Python 3.7
Python 3.3 has been EOL for over a year and current versions of some
tools (tox, virtualenv) now don't work with it anymore.  Instead of
keeping to install old versions we just remove it, while adding support
for Python 3.7 at the same time. The latter requires a workaround for
Travis for now.

The Docker images have been updated likewise and are now hosted on a
different account.  The updated Fedora package required some changes to
the expected output files.

Closes #49.